Fix crash when switching away from a dashboard with a choropleth chart
There's a bug in Immerse where it would crash if you clicked on the "Dashboards" link in the navbar while viewing a dashboard with a choropleth chart. The source of this appears to be mapd-charting sending a "remove legend" event in its `destroyChart()` function. Immerse will remove the chart from its state, then see the event and try to remove the legend from the already removed chart, leading to an error.

This commit makes it so the choropleth chart will not try to remove its legend when destroyed. This behavior is in line with the way other charts behave.