Fix formatting of color dimension
There was a, uh, bug?, where the color legend wasn't passing the color measure name to mapd3's `autoFormatter`. This meant that mapd3 wouldn't be able to associate the color legend values with a specific format string, and would default to the first format string found (for reasons). This commit changes it so that the color legend looks up the color measure name before formatting the legend values.