[FE-8934] Fix hpt binning with colors in histogram (#370)
* Use millisecond resolution for Dates in multi-series-mixin

In the `processMultiSeriesResults()` function in `multi-series-mixin.js`, add a check to see if the values be used are Date objects. If so, explicitly call `toISOString()` instead of letting it implicitly call `toString()`, so that we retain millisecond precision.

This change should fix using HPT values on histogram with a color dimension. Previously, you would only get one bin, since `toString()` cuts off milliseconds. Now, we get the proper binning.