[FE-9007] Handle invalid Date objects in `compareDomains()` (#375)
In certain circumstances (like creating a date global filter, and just clicking inside the start/end time inputs without changing them), the Date objects in `d2` (the second domain array) are invalid. Like, they're Date objects, but were created with invalid values. I'm not sure how or why these invalid Date objects are being created, but nonetheless, calling `toISOString()` on them throws an exception.

Due to JavaScript being JavaScript, calling `toString()` does not throw an exception; it just returns the string "Invalid date". So in `compareDomains()`, we test if the Date is valid (by checking `getTime()` doesn't return NaN). If it's valid, we call `toISOString()` like normal; if it's invalid, we just call `toString()` to avoid the exception.

This bug was introduced by PR #370 (FE-8934), which changed `toString()` to `toISOString()` in `compareDomains()`.