Fixes some tests.
NOTE: the change to the stacked-legend::should gradient test is
questionable. It's possible the changes made to the stacked-legend may
break legend behavior and the fact that this test is failing might be
and indicator of that. This may require additional investigation to
verify or not behavior changes. Going to land this in the meantime to
unblock and deal with any legend issues secondarily.