Adds windbarb raster chart layer via a new raster layer state parser/validator/vega-generator [DRAFT] (#585)
* Adds a new windbarb raster chart layer that aligns with it BE-
rendering counterpart. Vega-lite-like state json parsing is done
with a more stable parser/validator and vega generator.

Initial changes include:
 * adding a new exampleWindBarbs example with webpack config
 * adding a new windbarb raster layer called raster-layer-windarb-mixin
   * embedded in this file is a first iteration of a more refined raster
     layer state parser/validator and vega generator.
   * once this new parser/validator/generator is in a more solid state,
     the expectation is to change the other existing raster charts to
     it.
 * exposing the new windbarb raster layer in raster-chart, raster-layer
   and index

* Adds initial implementations of other scales that haven't been fully
utilized, namely threshold scale. The scales come with default "auto"
handling. For example, threshold scales use sigma values as stops
for "auto"

* Adds more formal legend support, including options to build legends
for threshold and quantize scales more appropriately.

* Fixes some tests to account for changes in stacked-legend

NOTE: the change to the stacked-legend::should gradient test is
questionable. It's possible the changes made to the stacked-legend may
break legend behavior and the fact that this test is failing might be
and indicator of that. This may require additional investigation to
verify or not behavior changes. Going to land this in the meantime to
unblock and deal with any legend issues secondarily.